+++ backend/src/gateway/chat.gateway.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { ChatGateway } from './chat.gateway';
+
+describe('ChatGateway', () => {
+	let gateway: ChatGateway;
+
+	beforeEach(async () => {
+		const module: TestingModule = await Test.createTestingModule({
+			providers: [ChatGateway],
+		}).compile();
+
+		gateway = module.get<ChatGateway>(ChatGateway);
+	});
+
+	it('should be defined', () => {
+		expect(gateway).toBeDefined();
+	});
+});
